OAK-11832 : moved Traverser to internal.graph package and marked as internal
